Add files via upload
##Description
This is a small program I worked on while I was practicing loops in Python. I wrote a program that would calculate test averages for each student and also a class average based on the number of students and number of tests. Feel free to run the code and give it a try.